Split out reducers from web flow
